+++ src/main/java/cloud/metaapi/sdk/meta_api/MetaApi.java
@@ -7,6 +7,7 @@ import org.apache.log4j.Logger;
 import cloud.metaapi.sdk.clients.HttpClient;
 import cloud.metaapi.sdk.clients.RetryOptions;
 import cloud.metaapi.sdk.clients.error_handler.ValidationException;
+import cloud.metaapi.sdk.clients.meta_api.HistoricalMarketDataClient;
 import cloud.metaapi.sdk.clients.meta_api.MetaApiWebsocketClient;
 import cloud.metaapi.sdk.clients.meta_api.MetatraderAccountClient;
 import cloud.metaapi.sdk.clients.meta_api.MetatraderDemoAccountClient;
@@ -70,6 +71,10 @@ public class MetaApi {
      * Retry options
      */
     public RetryOptions retryOpts = new RetryOptions();
+    /**
+     * Historical market data request timeout in milliseconds
+     */
+    public int historicalMarketDataRequestTimeout;
   }
   
   /**
@@ -142,6 +147,8 @@ public class MetaApi {
       throw new ValidationException("Application name must be non-empty string consisting from letters, digits and _ only", null);
     }
     HttpClient httpClient = new HttpClient(opts.requestTimeout * 1000, opts.connectTimeout * 1000, opts.retryOpts);
+    HttpClient historicalMarketDataHttpClient = new HttpClient(opts.historicalMarketDataRequestTimeout,
+      opts.connectTimeout * 1000, opts.retryOpts);
     MetaApiWebsocketClient.ClientOptions websocketOptions = new MetaApiWebsocketClient.ClientOptions();
     websocketOptions.application = opts.application;
     websocketOptions.domain = opts.domain;
@@ -154,8 +161,10 @@ public class MetaApi {
     metaApiWebsocketClient = new MetaApiWebsocketClient(token, websocketOptions);
     provisioningProfileApi = new ProvisioningProfileApi(new ProvisioningProfileClient(httpClient, token, opts.domain));
     connectionRegistry = new ConnectionRegistry(metaApiWebsocketClient, opts.application);
+    HistoricalMarketDataClient historicalMarketDataClient = new HistoricalMarketDataClient(
+      historicalMarketDataHttpClient, token, opts.domain);
     metatraderAccountApi = new MetatraderAccountApi(new MetatraderAccountClient(httpClient, token, opts.domain),
-      metaApiWebsocketClient, connectionRegistry);
+      metaApiWebsocketClient, connectionRegistry, historicalMarketDataClient);
     metatraderDemoAccountApi = new MetatraderDemoAccountApi(
       new MetatraderDemoAccountClient(httpClient, token, opts.domain));
     if (opts.enableLatencyMonitor) {