+++ src/main/java/cloud/metaapi/sdk/meta_api/MetatraderAccountApi.java
@@ -5,6 +5,7 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 import java.util.stream.Collectors;
 
+import cloud.metaapi.sdk.clients.meta_api.HistoricalMarketDataClient;
 import cloud.metaapi.sdk.clients.meta_api.MetaApiWebsocketClient;
 import cloud.metaapi.sdk.clients.meta_api.MetatraderAccountClient;
 import cloud.metaapi.sdk.clients.meta_api.models.AccountsFilter;
@@ -18,6 +19,7 @@ public class MetatraderAccountApi {
   private MetatraderAccountClient metatraderAccountClient;
   private MetaApiWebsocketClient metaApiWebsocketClient;
   private ConnectionRegistry connectionRegistry;
+  private HistoricalMarketDataClient historicalMarketDataClient;
   
   /**
    * Constructs a MetaTrader account API instance
@@ -27,10 +29,12 @@ public class MetatraderAccountApi {
    * @param expertAdvisorClient expert advisor REST API client
    */
   public MetatraderAccountApi(MetatraderAccountClient metatraderAccountClient,
-    MetaApiWebsocketClient metaApiWebsocketClient, ConnectionRegistry connectionRegistry) {
+    MetaApiWebsocketClient metaApiWebsocketClient, ConnectionRegistry connectionRegistry,
+    HistoricalMarketDataClient historicalMarketDataClient) {
     this.metatraderAccountClient = metatraderAccountClient;
     this.metaApiWebsocketClient = metaApiWebsocketClient;
     this.connectionRegistry = connectionRegistry;
+    this.historicalMarketDataClient = historicalMarketDataClient;
   }
   
   /**
@@ -49,7 +53,7 @@ public class MetatraderAccountApi {
   public CompletableFuture<List<MetatraderAccount>> getAccounts(AccountsFilter accountsFilter) {
     return metatraderAccountClient.getAccounts(accountsFilter).thenApply(accounts -> {
       return accounts.stream().map(accountDto -> new MetatraderAccount(accountDto, metatraderAccountClient,
-        metaApiWebsocketClient, connectionRegistry)).collect(Collectors.toList());
+        metaApiWebsocketClient, connectionRegistry, historicalMarketDataClient)).collect(Collectors.toList());
     });
   }
   
@@ -61,7 +65,7 @@ public class MetatraderAccountApi {
   public CompletableFuture<MetatraderAccount> getAccount(String accountId) {
     return metatraderAccountClient.getAccount(accountId).thenApply(accountDto -> {
       return new MetatraderAccount(accountDto, metatraderAccountClient, metaApiWebsocketClient,
-        connectionRegistry);
+        connectionRegistry, historicalMarketDataClient);
     });
   }
   
@@ -72,7 +76,7 @@ public class MetatraderAccountApi {
   public CompletableFuture<MetatraderAccount> getAccountByToken() {
     return metatraderAccountClient.getAccountByToken().thenApply(accountDto -> {
       return new MetatraderAccount(accountDto, metatraderAccountClient, metaApiWebsocketClient,
-        connectionRegistry);
+        connectionRegistry, historicalMarketDataClient);
     });
   }